+++ src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java
@@ -4,7 +4,7 @@ import ru.yandex.practicum.filmorate.model.*;
 import java.util.Collection;
 
 public interface FilmService {
-    Film getFilmById(Integer id);
+    Film getFilmById(Long id);
 
     Collection<Film> getAllFilms();
 
@@ -22,9 +22,9 @@ public interface FilmService {
 
     Film updateFilm(Film film);
 
-    Film putLike(Integer id, Long userId);
+    Film putLike(Long id, Long userId);
 
-    Film deleteLike(Integer id, Long userId);
+    Film deleteLike(Long id, Long userId);
 
-    String deleteFilm(Integer id);
+    String deleteFilm(Long id);
 }
